+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Read an overview of the information in the headers of a given message or sequence 
 - Folder name encoding added (#324)
 - Add a fallback when aliasing encodings for iconv (#325)
+- Method to receive the encoded folder name added
 
 ### Breaking changes
 - NaN